+++ src/circle_fighter/game/plane/bounds/PlaneBounds.java
@@ -4,9 +4,9 @@ import circle_fighter.game.object.GameObject;
 import circle_fighter.game.object.position.Position;
 
 public abstract class PlaneBounds{
-    public abstract double exceedsLeftBy(GameObject object);
-    public abstract double exceedsRightBy(GameObject object);
-    public abstract double exceedsTopBy(GameObject object);
-    public abstract double exceedsBottomBy(GameObject object);
+    public abstract float exceedsLeftBy(GameObject object);
+    public abstract float exceedsRightBy(GameObject object);
+    public abstract float exceedsTopBy(GameObject object);
+    public abstract float exceedsBottomBy(GameObject object);
     public abstract Position getPositionOffset(Position targetOffset);
 }
